Implement variable toggling function
Wrote the function toggle-variable, which toggles its argument between
t and nil. In particular, implemented this to easily toggle the
emphasis markers on and off in org-mode, so set a keybinding for that.

Also added a latex snippet for rightarrow macros and made use-package
load everything immediately if it's running in a daemon.